backend/requirements.txt
python-dotenv==1.0.1
openai==1.14.1
pytest==8.0.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
zipp==3.17.0
numpy
pandas
scipy
flask-mysqldb
flask_cors
pyaudio
